Made Rotatable ABC. However, it seems that my usage of ABC doesn't enforce implementation of abstractmethod. I swear it worked when I tried it out. We might want to look into this at some point.
